+++ src/app/modules/modals/send-solution-modal/send-solution-modal.component.ts
@@ -5,6 +5,7 @@ import { ChallengeService } from '../../../services/challenge.service';
 import { AuthService } from '../../../services/auth.service';
 import { SolutionStatus } from 'src/app/models/user-solution-status.enum';
 import { ChallengeTab } from 'src/app/shared/enums/challenge-tab.enum';
+import { SubmitSolutionResponse } from 'src/app/models/user-solution.interface';
 
 @Component({
   selector: 'app-send-solution-modal',
@@ -23,7 +24,7 @@ export class SendSolutionModalComponent {
   solutionText: string = '';
   @Output() solutionAccepted = new EventEmitter<boolean>();
   @Output() solutionSubmitted = new EventEmitter<string>();
-
+  @Output() timesSolvedUpdated = new EventEmitter<number>();
   ngOnInit(): void {
     this.getLanguageId(); 
     this.getSolutionText(); 
@@ -52,7 +53,8 @@ export class SendSolutionModalComponent {
       SolutionStatus.ENDED,
       this.solutionText
     ).subscribe({
-      next: (response) => {
+      next: (response: SubmitSolutionResponse) => {
+        this.timesSolvedUpdated.emit(response.timesSolved);
         const solutionText = response.solution_text;
         this.solutionService.solutionText(solutionText);
         this.solutionSubmitted.emit(solutionText);