+++ src/app/services/solution.service.ts
@@ -1,9 +1,9 @@
 import { HttpClient } from '@angular/common/http'
 import { Injectable, inject } from '@angular/core'
-import { BehaviorSubject, of, Subject, type Observable } from 'rxjs'
+import { BehaviorSubject, catchError, map, of, Subject, type Observable } from 'rxjs'
 import { environment } from 'src/environments/environment'
 import { type DataSolution } from '../models/data-solution.model'
-import { type UserSolution } from '../models/user-solution.interface'
+import { SubmitSolutionResponse, type UserSolution } from '../models/user-solution.interface'
 import { ChallengeTab } from 'src/app/shared/enums/challenge-tab.enum'
 
 @Injectable({
@@ -78,6 +78,14 @@ export class SolutionService {
           'Content-Type': 'application/json'
         }
       }
+    ).pipe(
+      map((response: SubmitSolutionResponse) => {
+        return {
+          ...response,
+          isSolved: true,
+          timesSolved: response.timesSolved ?? 1
+        }
+      })
     );
   }